Refactor project structure, update imports, and improve type safety
