+++ components/ui/context-menu.tsx
@@ -4,7 +4,7 @@ import * as React from "react"
 import * as ContextMenuPrimitive from "@radix-ui/react-context-menu"
 import { Check, ChevronRight, Circle } from "lucide-react"
 
-import { cn } from "@lib/utils"
+import { cn } from "@/lib/utils"
 
 const ContextMenu = ContextMenuPrimitive.Root
